requirements.txt
pydantic~=2.0.0
requests~=2.31.0
RestrictedPython~=6.0
PyYAML~=6.0
flake8~=6.0.0
python-dotenv~=1.0.0 
openai~=1.3.0 